third_party/WebKit/LayoutTests/external/wpt/fetch/api/redirect/redirect-referrer-worker-expected.txt

This is a testharness.js-based test.
PASS Same origin redirection, empty init, unsafe-url redirect header  
PASS Same origin redirection, empty init, no-referrer-when-downgrade redirect header  
PASS Same origin redirection, empty init, same-origin redirect header  
PASS Same origin redirection, empty init, origin redirect header  
PASS Same origin redirection, empty init, origin-when-cross-origin redirect header  
PASS Same origin redirection, empty init, no-referrer redirect header  
FAIL Same origin redirection, empty init, strict-origin redirect header  assert_equals: Check referrer header expected "http://web-platform.test:8001/" but got "http://web-platform.test:8001/fetch/api/redirect/redirect-referrer.js"
PASS Same origin redirection, empty init, strict-origin-when-cross-origin redirect header  
PASS Same origin redirection, empty redirect header, unsafe-url init  
PASS Same origin redirection, empty redirect header, no-referrer-when-downgrade init  
FAIL Same origin redirection, empty redirect header, same-origin init  promise_test: Unhandled rejection with value: object "TypeError: Failed to execute 'fetch' on 'WorkerGlobalScope': Invalid referrer policy"
PASS Same origin redirection, empty redirect header, origin init  
PASS Same origin redirection, empty redirect header, origin-when-cross-origin init  
PASS Same origin redirection, empty redirect header, no-referrer init  
FAIL Same origin redirection, empty redirect header, strict-origin init  promise_test: Unhandled rejection with value: object "TypeError: Failed to execute 'fetch' on 'WorkerGlobalScope': Invalid referrer policy"
FAIL Same origin redirection, empty redirect header, strict-origin-when-cross-origin init  promise_test: Unhandled rejection with value: object "TypeError: Failed to execute 'fetch' on 'WorkerGlobalScope': Invalid referrer policy"
PASS Cross origin redirection, empty init, unsafe-url redirect header  
PASS Cross origin redirection, empty init, no-referrer-when-downgrade redirect header  
FAIL Cross origin redirection, empty init, same-origin redirect header  assert_equals: Check referrer header expected (object) null but got (string) "http://web-platform.test:8001/fetch/api/redirect/redirect-referrer.js"
PASS Cross origin redirection, empty init, origin redirect header  
PASS Cross origin redirection, empty init, origin-when-cross-origin redirect header  
PASS Cross origin redirection, empty init, no-referrer redirect header  
FAIL Cross origin redirection, empty init, strict-origin redirect header  assert_equals: Check referrer header expected "http://web-platform.test:8001/" but got "http://web-platform.test:8001/fetch/api/redirect/redirect-referrer.js"
FAIL Cross origin redirection, empty init, strict-origin-when-cross-origin redirect header  assert_equals: Check referrer header expected "http://web-platform.test:8001/" but got "http://web-platform.test:8001/fetch/api/redirect/redirect-referrer.js"
PASS Cross origin redirection, empty redirect header, unsafe-url init  
PASS Cross origin redirection, empty redirect header, no-referrer-when-downgrade init  
FAIL Cross origin redirection, empty redirect header, same-origin init  promise_test: Unhandled rejection with value: object "TypeError: Failed to execute 'fetch' on 'WorkerGlobalScope': Invalid referrer policy"
PASS Cross origin redirection, empty redirect header, origin init  
PASS Cross origin redirection, empty redirect header, origin-when-cross-origin init  
PASS Cross origin redirection, empty redirect header, no-referrer init  
FAIL Cross origin redirection, empty redirect header, strict-origin init  promise_test: Unhandled rejection with value: object "TypeError: Failed to execute 'fetch' on 'WorkerGlobalScope': Invalid referrer policy"
FAIL Cross origin redirection, empty redirect header, strict-origin-when-cross-origin init  promise_test: Unhandled rejection with value: object "TypeError: Failed to execute 'fetch' on 'WorkerGlobalScope': Invalid referrer policy"
Harness: the test ran to completion.
